Rename proc macro server from 'Rustc' to 'RustAnalyzer'
